+++ js/utils.js
@@ -1,0 +1,12 @@
+const createIdGenerator = () => {
+  let generatorId = 0;
+  return function() {
+    return ++generatorId;
+  };
+};
+
+const getRandomInteger = (min, max) => Math.floor(Math.random() * (max - min + 1)) + min;
+
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+export { createIdGenerator, getRandomInteger, getRandomArrayElement };
